packages/common_lib/pyproject.toml
[project]
name = "common-lib"
version = "0.1.0"
description = "A shared package"
authors = [
    {name = "Christian Cardin",email = "christian.cardin@iceye.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "common_lib", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.14.10,<0.15.0)"
]
